ctdb-protocol: Avoid signed/unsigned comparison by casting

+++ ctdb/protocol/protocol_types.c
@@ -1680,19 +1680,19 @@ int ctdb_rec_buffer_write(struct ctdb_rec_buffer *recbuf, int fd)
 	ssize_t n;
 
 	n = write(fd, &recbuf->db_id, sizeof(uint32_t));
-	if (n == -1 || n != sizeof(uint32_t)) {
+	if (n == -1 || (size_t)n != sizeof(uint32_t)) {
 		return (errno != 0 ? errno : EIO);
 	}
 	n = write(fd, &recbuf->count, sizeof(uint32_t));
-	if (n == -1 || n != sizeof(uint32_t)) {
+	if (n == -1 || (size_t)n != sizeof(uint32_t)) {
 		return (errno != 0 ? errno : EIO);
 	}
 	n = write(fd, &recbuf->buflen, sizeof(size_t));
-	if (n == -1 || n != sizeof(size_t)) {
+	if (n == -1 || (size_t)n != sizeof(size_t)) {
 		return (errno != 0 ? errno : EIO);
 	}
 	n = write(fd, recbuf->buf, recbuf->buflen);
-	if (n == -1 || n != recbuf->buflen) {
+	if (n == -1 || (size_t)n != recbuf->buflen) {
 		return (errno != 0 ? errno : EIO);
 	}
 
@@ -1711,15 +1711,15 @@ int ctdb_rec_buffer_read(int fd, TALLOC_CTX *mem_ctx,
 	}
 
 	n = read(fd, &recbuf->db_id, sizeof(uint32_t));
-	if (n == -1 || n != sizeof(uint32_t)) {
+	if (n == -1 || (size_t)n != sizeof(uint32_t)) {
 		return (errno != 0 ? errno : EIO);
 	}
 	n = read(fd, &recbuf->count, sizeof(uint32_t));
-	if (n == -1 || n != sizeof(uint32_t)) {
+	if (n == -1 || (size_t)n != sizeof(uint32_t)) {
 		return (errno != 0 ? errno : EIO);
 	}
 	n = read(fd, &recbuf->buflen, sizeof(size_t));
-	if (n == -1 || n != sizeof(size_t)) {
+	if (n == -1 || (size_t)n != sizeof(size_t)) {
 		return (errno != 0 ? errno : EIO);
 	}
 
@@ -1729,7 +1729,7 @@ int ctdb_rec_buffer_read(int fd, TALLOC_CTX *mem_ctx,
 	}
 
 	n = read(fd, recbuf->buf, recbuf->buflen);
-	if (n == -1 || n != recbuf->buflen) {
+	if (n == -1 || (size_t)n != recbuf->buflen) {
 		return (errno != 0 ? errno : EIO);
 	}
